feat(#656): [Pooja] refactor to get language details from i18n
